[Rebalance] Change default give up/takeover routine count to 1
* Give up routine count control the max no. of vbuckets whose ownership
  will be given up at any given point in time during rebalance.
  Similarly takeover routine count controls the max no. of vbuckets
  whose ownership will be requested by a given node at any given point
  in time. They basically control how soon rebalance could finish at the
  expense of lowering down throughput of event processing for the
  vbuckets in flight. Changing the default count of those routines to
  1 and it could be tuned using rest calls at the time of app deploy.

* Plasma entries that were migrated to another eventing node weren't
  getting cleaned up on source node. Change addresses that bug.

Reviewed-on: http://review.couchbase.org/84868
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ consumer/timer_events.go
@@ -150,21 +150,6 @@ func (r *timerProcessingWorker) processTimerEvents() {
 		select {
 		case <-r.stopCh:
 			return
-		case vb := <-r.signalProcessTimerPlasmaCloseCh:
-			// Rebalance takeover routine will send signal on this channel to signify
-			// stopping of any plasma.Writer instance for a specific vbucket
-			r.c.plasmaReaderRWMutex.Lock()
-			_, ok := r.c.vbPlasmaReader[vb]
-			if ok {
-				delete(r.c.vbPlasmaReader, vb)
-			}
-
-			r.c.plasmaReaderRWMutex.Unlock()
-
-			// sends ack message back to rebalance takeover routine, so that it could
-			// safely call Close() on vb specific plasma store
-			r.c.signalProcessTimerPlasmaCloseAckCh <- vb
-			continue
 		case <-r.timerProcessingTicker.C:
 		}
 
@@ -338,8 +323,10 @@ func (c *Consumer) processNonDocTimerEvents(cTimer, nTimer string, bootstrap boo
 							logging.Errorf("CRTE[%s:%s:%s:%d] vb: %v Cron timer key: %v val: %v, err: %v",
 								c.app.AppName, c.workerName, c.tcpPort, c.Pid(), vb, currTimer, val, err)
 						}
-						c.nonDocTimerEntryCh <- timerMsg{payload: string(data), msgCount: len(val.CronTimers)}
-						c.gocbMetaBucket.Remove(currTimer, 0)
+						if len(val.CronTimers) > 0 {
+							c.nonDocTimerEntryCh <- timerMsg{payload: string(data), msgCount: len(val.CronTimers)}
+							c.gocbMetaBucket.Remove(currTimer, 0)
+						}
 					}
 					c.updateNonDocTimerStats(vb)
 				}
